Enable cr status/condition update after kong configuration
